pi/rover-control/requirements.txt
lxml==4.9.0
parse==1.19.0
pyserial==3.5
